requirements.txt
umap-learn
seaborn
openpyxl
#sklearn
requests

numpy
matplotlib